refactoring to reuse code between dataset and datastream translation
