+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/plan/schema/DataSetTable.scala
@@ -23,58 +23,15 @@ import java.util
 import java.util.Collections
 
 import org.apache.calcite.rel.{RelCollation, RelDistribution}
-import org.apache.calcite.rel.`type`.{RelDataType, RelDataTypeFactory}
 import org.apache.calcite.schema.Statistic
-import org.apache.calcite.schema.impl.AbstractTable
-import org.apache.calcite.sql.`type`.SqlTypeName
 import org.apache.calcite.util.ImmutableBitSet
-import org.apache.flink.api.common.typeinfo.AtomicType
-import org.apache.flink.api.common.typeutils.CompositeType
 import org.apache.flink.api.java.DataSet
-import org.apache.flink.api.table.typeutils.TypeConverter
 
 class DataSetTable[T](
     val dataSet: DataSet[T],
-    val fieldIndexes: Array[Int],
-    val fieldNames: Array[String])
-  extends AbstractTable {
-
-  if (fieldIndexes.length != fieldNames.length) {
-    throw new IllegalArgumentException(
-      "Number of field indexes and field names must be equal.")
-  }
-
-  // check uniquenss of field names
-  if (fieldNames.length != fieldNames.toSet.size) {
-    throw new IllegalArgumentException(
-      "Table field names must be unique.")
-  }
-
-  val fieldTypes: Array[SqlTypeName] =
-    dataSet.getType match {
-      case cType: CompositeType[T] =>
-        if (fieldNames.length != cType.getArity) {
-          throw new IllegalArgumentException(
-          s"Arity of DataSet type (" + cType.getFieldNames.deep + ") " +
-            "not equal to number of field names " + fieldNames.deep + ".")
-        }
-        fieldIndexes
-          .map(cType.getTypeAt(_))
-          .map(TypeConverter.typeInfoToSqlType(_))
-      case aType: AtomicType[T] =>
-        if (fieldIndexes.length != 1 || fieldIndexes(0) != 0) {
-          throw new IllegalArgumentException(
-            "Non-composite input type may have only a single field and its index must be 0.")
-        }
-        Array(TypeConverter.typeInfoToSqlType(aType))
-    }
-
-  override def getRowType(typeFactory: RelDataTypeFactory): RelDataType = {
-    val builder = typeFactory.builder
-    fieldNames.zip(fieldTypes)
-      .foreach( f => builder.add(f._1, f._2).nullable(true) )
-    builder.build
-  }
+    override val fieldIndexes: Array[Int],
+    override val fieldNames: Array[String])
+  extends FlinkTable[T](dataSet.getType, fieldIndexes, fieldNames) {
 
   override def getStatistic: Statistic = {
     new DefaultDataSetStatistic
